Commit all changes (untested).

+++ cctools-773/as/read.c
@@ -2526,7 +2526,7 @@ uintptr_t value)
 	*p = 0;
 	if(strlen(segname) > sizeof(s.segname)){
 	    as_bad("segment-name: %s too long (maximum %ld characters)",
-		    segname, sizeof(s.segname));
+		    segname, (long int) sizeof(s.segname));
 	    ignore_rest_of_line();
 	    *p = c;
 	    return;
@@ -2535,7 +2535,7 @@ uintptr_t value)
 	*q = 0;
 	if(strlen(sectname) > sizeof(s.sectname)){
 	    as_bad("section-name: %s too long (maximum %ld characters)",
-		    sectname, sizeof(s.sectname));
+		    sectname, (long int) sizeof(s.sectname));
 	    ignore_rest_of_line();
 	    return;
 	}
@@ -2689,7 +2689,7 @@ uintptr_t value)
 	*p = 0;
 	if(strlen(segname) > sizeof(s.segname)){
 	    as_bad("segment-name: %s too long (maximum %ld characters)",
-		    segname, sizeof(s.segname));
+		    segname, (long int) sizeof(s.segname));
 	    ignore_rest_of_line();
 	    *p = c;
 	    return;
@@ -2698,7 +2698,7 @@ uintptr_t value)
 	*q = 0;
 	if(strlen(sectname) > sizeof(s.sectname)){
 	    as_bad("section-name: %s too long (maximum %ld characters)",
-		    sectname, sizeof(s.sectname));
+		    sectname, (long int) sizeof(s.sectname));
 	    ignore_rest_of_line();
 	    *p = c;
 	    *q = d;
